requirements.txt
asgiref==3.8.1
blinker==1.4
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.1
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.4
django-allauth==65.3.1
djangorestframework==3.15.2
httplib2==0.20.2
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
mutagen==1.45.1
netifaces==0.11.0
numpy==2.1.1
oauthlib==3.2.0
pandas==2.2.3
psutil==5.9.0
pycryptodomex==3.11.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
PyMySQL==1.1.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytz==2024.2
pyxattr==0.7.2
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.2
ssh-import-id==5.11
systemd-python==234
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
urwid==2.1.2
wadllib==1.3.6
websockets==9.1
yt-dlp==2022.4.8
zipp==1.0.0
